pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lkml-bot"
version = "0.1.0"
description = "LKML Bot - A mailing list monitoring bot for Linux kernel and other subsystems"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.0.0",  # 包含 FastAPI driver 依赖
    "nonebot-adapter-discord>=0.1.8",
    "nonebot-adapter-feishu>=2.6.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",              # HTTP 客户端
    "websockets>=11.0.0",         # WebSocket 支持
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]


[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.nonebot]
adapters = [
    { name = "飞书", module_name = "nonebot.adapters.feishu" },
    { name = "Discord", module_name = "nonebot.adapters.discord" }
]
plugins = []
plugin_dirs = ["src/plugins"]
builtin_plugins = []

# Ruff 代码检查和格式化配置
[tool.ruff]
target-version = "py39"
line-length = 100

# 基础规则集（后续可逐步添加）
select = [
    "E",   # pycodestyle errors（基础错误）
    "W",   # pycodestyle warnings（基础警告）
    "F",   # Pyflakes（未使用的导入、变量等）
    "I",   # isort（导入排序）
]

# 忽略的规则
ignore = [
    "E501",  # line too long（由格式化工具处理）
]

# 排除的目录
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.isort]
known-first-party = ["lkml", "lkml_bot"]

# MyPy 类型检查配置（基础配置）
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
show_error_codes = true

# 忽略缺少类型注解的导入
[[tool.mypy.overrides]]
module = [
    "feedparser.*",
    "nonebot.*",
    "nonebot.adapters.*",
]
ignore_missing_imports = true
